requirements.txt
httpx
google-genai
kokoro>=0.9
torch
soundfile
numpy
pydub
pyyaml
python-dotenv
click
python-telegram-bot
fastapi
uvicorn[standard]
jinja2
sqlmodel
python-multipart
itsdangerous
